requirements.txt
python-dotenv
pydantic
langchain==1.0.1
langchain-core
langchain-chroma
langchain-community
langchain-text-splitters
langchain-google-genai
langchain-huggingface
sentence-transformers
chromadb
gradio